move <CommonAvatar /> to ui-kit
