+++ src/pages/common/components/CommonHeader/CommonHeader.tsx
@@ -3,7 +3,7 @@ import classNames from "classnames";
 import { useCommonDataContext } from "@/pages/common/providers/CommonData/context";
 import { useIsTabletView } from "@/shared/hooks/viewport";
 import { Button, ButtonSize, ButtonVariant } from "@/shared/ui-kit/Button";
-import { CommonAvatar } from "../CommonAvatar";
+import { CommonAvatar } from "@/shared/ui-kit/CommonAvatar";
 import { KeyValueItem, KeyValuePairs } from "./components";
 import styles from "./CommonHeader.module.scss";
