+++ src/shared/ui-kit/CommonAvatar/CommonAvatar.tsx
