+++ src/shared/ui-kit/CommonAvatar/index.ts
